XDPXCONF-113 Need an API for Xconf Admin Telemetry profile (so that team can add / remove markers via API): - fixed pending changes validation - moved classed to data-service module
