+++ xconf-angular-admin/src/test/java/com/comcast/xconf/admin/controller/change/ChangeControllerTest.java
@@ -25,8 +25,6 @@ import com.comcast.apps.dataaccess.util.JsonUtil;
 import com.comcast.hydra.astyanax.data.IPersistable;
 import com.comcast.xconf.admin.controller.BaseControllerTest;
 import com.comcast.xconf.admin.controller.telemetry.PermanentProfileController;
-import com.comcast.xconf.admin.service.change.ApprovedChangeCrudService;
-import com.comcast.xconf.admin.service.change.ChangeCrudService;
 import com.comcast.xconf.admin.service.change.TelemetryProfileChangeService;
 import com.comcast.xconf.admin.service.telemetry.PermanentTelemetryProfileService;
 import com.comcast.xconf.change.ApprovedChange;
@@ -34,6 +32,8 @@ import com.comcast.xconf.change.Change;
 import com.comcast.xconf.logupload.telemetry.PermanentTelemetryProfile;
 import com.comcast.xconf.logupload.telemetry.TelemetryProfile;
 import com.comcast.xconf.search.SearchFields;
+import com.comcast.xconf.service.change.ApprovedChangeCrudService;
+import com.comcast.xconf.service.change.ChangeCrudService;
 import com.google.common.collect.Lists;
 import org.junit.After;
 import org.junit.Before;
