+++ xconf-angular-admin/src/test/java/com/comcast/xconf/admin/controller/change/TelemetryTwoChangeControllerTest.java
@@ -25,13 +25,13 @@ import com.comcast.hydra.astyanax.data.IPersistable;
 import com.comcast.xconf.admin.controller.BaseControllerTest;
 import com.comcast.xconf.admin.controller.telemetry.TelemetryTwoProfileController;
 import com.comcast.xconf.admin.service.telemetry.TelemetryTwoProfileService;
-import com.comcast.xconf.admin.service.telemetrytwochange.ApprovedTelemetryTwoChangeCrudService;
-import com.comcast.xconf.admin.service.telemetrytwochange.TelemetryTwoChangeCrudService;
 import com.comcast.xconf.admin.service.telemetrytwochange.TelemetryTwoProfileChangeService;
 import com.comcast.xconf.change.ApprovedTelemetryTwoChange;
 import com.comcast.xconf.change.TelemetryTwoChange;
 import com.comcast.xconf.logupload.telemetry.TelemetryTwoProfile;
 import com.comcast.xconf.search.SearchFields;
+import com.comcast.xconf.service.telemetrytwochange.ApprovedTelemetryTwoChangeCrudService;
+import com.comcast.xconf.service.telemetrytwochange.TelemetryTwoChangeCrudService;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
