+++ xconf-angular-admin/src/test/java/com/comcast/xconf/admin/controller/telemetry/PermanentProfileControllerTest.java
@@ -23,14 +23,14 @@ import com.comcast.apps.dataaccess.util.CloneUtil;
 import com.comcast.apps.dataaccess.util.JsonUtil;
 import com.comcast.hydra.astyanax.data.IPersistable;
 import com.comcast.xconf.admin.controller.BaseControllerTest;
-import com.comcast.xconf.admin.service.change.ApprovedChangeCrudService;
-import com.comcast.xconf.admin.service.change.ChangeCrudService;
 import com.comcast.xconf.change.ApprovedChange;
 import com.comcast.xconf.change.Change;
 import com.comcast.xconf.change.ChangeOperation;
 import com.comcast.xconf.firmware.ApplicationType;
 import com.comcast.xconf.logupload.telemetry.PermanentTelemetryProfile;
 import com.comcast.xconf.search.SearchFields;
+import com.comcast.xconf.service.change.ApprovedChangeCrudService;
+import com.comcast.xconf.service.change.ChangeCrudService;
 import com.google.common.collect.Lists;
 import org.junit.After;
 import org.junit.Before;
