+++ xconf-angular-admin/src/test/java/com/comcast/xconf/admin/controller/telemetry/TelemetryTwoProfileControllerTest.java
@@ -22,13 +22,13 @@ import com.comcast.apps.dataaccess.dao.SimpleDao;
 import com.comcast.apps.dataaccess.util.JsonUtil;
 import com.comcast.hydra.astyanax.data.IPersistable;
 import com.comcast.xconf.admin.controller.BaseControllerTest;
-import com.comcast.xconf.admin.service.telemetrytwochange.TelemetryTwoChangeCrudService;
 import com.comcast.xconf.change.ApprovedTelemetryTwoChange;
 import com.comcast.xconf.change.TelemetryTwoChange;
 import com.comcast.xconf.firmware.ApplicationType;
 import com.comcast.xconf.logupload.telemetry.TelemetryTwoProfile;
 import com.comcast.xconf.logupload.telemetry.TelemetryTwoRule;
 import com.comcast.xconf.search.SearchFields;
+import com.comcast.xconf.service.telemetrytwochange.TelemetryTwoChangeCrudService;
 import com.google.common.collect.Lists;
 import org.junit.After;
 import org.junit.Before;
