+++ xconf-dataservice/src/test/java/com/comcast/xconf/CompleteTestSuite.java
@@ -42,8 +42,6 @@ import org.junit.BeforeClass;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
-import java.io.IOException;
-
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
         DCMTestCaseFrom1To3.class, DCMTestCaseFrom4To7.class,
@@ -89,7 +87,7 @@ import java.io.IOException;
         GenericNamespacedListQueriesServiceTest.class,
         FeatureDataControllerTest.class, FeatureRuleDataControllerTest.class,
         ActivationVersionDataControllerTest.class,
-        TelemetryProfileControllerTest.class, TelemetryTwoDataControllerTest.class
+        TelemetryProfileDataControllerTest.class, TelemetryProfileTwoDataControllerTest.class
 })
 public class CompleteTestSuite {
 
@@ -97,7 +95,7 @@ public class CompleteTestSuite {
     public static final long timeToWaitExpiration = telemetryProfileServiceExpireTimeMs + 500L;
 
     @BeforeClass
-    public static void setUpProperties() throws IOException {
+    public static void setUpProperties() {
         TelemetryProfileService.expireTime = telemetryProfileServiceExpireTimeMs;
     }
 }
